semi-supervised-node-classification/requirements.txt
lmdb
torch-scatter
torch-sparse
torch-geometric
tensorboardX
tensorboard
